Use Stack ID for OriginAccessControl.Name property (#69)
* Add link to repo in template description
* Derive OACConfig name from StackId
* Bump version

Use the Stack ID for the `OriginAccessControlConfig.Name` property.

The template fails to deploy when the stack name is longer than 64 characters. The `OriginAccessControlConfig` resource fails to create, with a generic `InvalidRequest` error. The `Name` property has. a max length of 64 characters.

The OriginAccessControlConfig.Name property was set to `!Sub 'oac-${AWS::StackName}-${AWS::Region}'`. For sufficiently long stack names (including the default stack name), this triggered the bug described above.

`OriginAccessControlConfig` resources must have a unique `Name` across all regions for an account. To ensure this, the `Name` property was derived from both the StackName and Region. The max length of a stack name is 128 characters. This is longer than the 64 characters allowed by the OACConfig Name property.